+++ packages/opentelemetry-instrumentation-openai_agents/opentelemetry/instrumentation/openai_agents/utils.py
@@ -1,0 +1,16 @@
+import os
+from opentelemetry import context as context_api
+
+
+def set_span_attribute(span, name, value):
+    if value is not None:
+        if value != "":
+            span.set_attribute(name, value)
+    return
+
+
+def should_send_prompts():
+    return (
+        os.getenv("TRACELOOP_TRACE_CONTENT") or "true"
+    ).lower() == "true" or \
+        context_api.get_value("override_enable_content_tracing")
